requirements.txt
prefect==2.14.15
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
pandas==2.2.2
numpy==2.0.1
nltk==3.9.1
plotly==5.23.0
dash==2.18.2
python-dotenv==1.0.1
langdetect==1.0.9
beautifulsoup4==4.12.3
feedparser==6.0.11
requests==2.32.3
